backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pdfplumber
sentence-transformers
faiss-cpu
openai
numpy
